refactor: 코드 계층 구조 변경
