+++ src/main/java/org/sophy/sophy/controller/AuthorController.java
@@ -13,8 +13,8 @@ import org.sophy.sophy.domain.dto.booktalk.response.BooktalkCreateResponseDto;
 import org.sophy.sophy.domain.dto.booktalk.response.BooktalkDeleteResponseDto;
 import org.sophy.sophy.domain.dto.mypage.MyPageBooktalkDto;
 import org.sophy.sophy.exception.SuccessStatus;
-import org.sophy.sophy.service.api.BooktalkService;
-import org.sophy.sophy.service.api.MemberService;
+import org.sophy.sophy.service.BooktalkService;
+import org.sophy.sophy.service.MemberService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.User;
