+++ src/main/java/org/sophy/sophy/controller/MemberController.java
@@ -10,7 +10,7 @@ import org.sophy.sophy.controller.dto.request.MemberAdditionalInfoDto;
 import org.sophy.sophy.domain.dto.mypage.MyPageDto;
 import org.sophy.sophy.domain.dto.mypage.MyInfoDto;
 import org.sophy.sophy.exception.SuccessStatus;
-import org.sophy.sophy.service.api.MemberService;
+import org.sophy.sophy.service.MemberService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.User;
