+++ src/main/java/org/sophy/sophy/domain/CompletedBooktalk.java
@@ -13,7 +13,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.sophy.sophy.domain.enumerate.BookCategory;
-import org.sophy.sophy.domain.other.AuditingTimeEntity;
+import org.sophy.sophy.domain.common.AuditingTimeEntity;
 
 @Entity
 @Getter
