+++ src/main/java/org/sophy/sophy/domain/Member.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import java.util.ArrayList;
 import java.util.List;
-import org.sophy.sophy.domain.other.AuditingTimeEntity;
+import org.sophy.sophy.domain.common.AuditingTimeEntity;
 
 @Entity
 @Getter
