+++ src/main/java/org/sophy/sophy/service/BooktalkService.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.service.api;
+package org.sophy.sophy.service;
 
 import java.util.ArrayList;
 import java.util.Comparator;
