+++ src/main/java/org/sophy/sophy/service/HomeService.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.service.api;
+package org.sophy.sophy.service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
