+++ src/main/java/org/sophy/sophy/service/MemberService.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.service.api;
+package org.sophy.sophy.service;
 
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
